Working prototype
Given a directory path as an argument (it shouldn't be mandatory), it
finds `.git` directories into the immediate path given. For every
repository found a `git fetch --all` is issued.

The idea is to be able to track development of multiple repositories,
from multiple developers (remotes) easily. It is good to have a summary
from now and than and it is awesome when developers who work with you
use their private repositories to develop ideas and expose them
publicly.

The idiotic name comes from asking Git "What is up": `git sup`!

Get it yo?!

Stuff that bothers me
---------------------

* It does not work if the directory you are on is a git repository.
  *shame* *shame* *shame*

* When a big `fetch` is issued, the "downloading" messages go to the
  output - probably stderr. They should not be there, I don't want to
  read how much I have downloaded.

* A path is always required as first argument. It should default - as
  Git - to the current path I am on.

* A lot of tests are missing for multiple repositories - all of them.

* Missing the "ignore remote" option usage, currently it just hides the
  output but the remote is fetched anyway. We should not fetch ignored
  remotes.
  As a result, `git fetch` should always be done to an specific remote and
  not with `--all` as a default option.