Supports help to `git-sup fetch`
One can add a help inside a sub command, it feels better then specifying
a different function to output the help of that sub command. Each sub
command might act differently in the future, so this will help keep
things together.

Reorganizes functions so they are better described (by renaming them)
and "private" functions are kept first. Private functions start with
"_".

Removes functionality of ignoring remotes, it was not properly
implemented or documented.