Allows listing unique authors from repositories
E-mail is preferred as the first column because is the least likely
to suffer from changes and typos, along with being the most important
information inside a `.mailmap` file. The Unique filter is not applied
solely in the e-mail though, but on the combination of e-mail and name -
which can cause the same e-mail to appear multiple times.

A section in the help of the command is added, describing the columns on
a line in the order they appear. This can be parsed later to allow
selection of custom columns with just one command parsing this.

A BREAKING CHNAGE on the _git wrapper is added, redirecting stderr to
/dev/null. Avoiding error messages popping out and being treated as
meaningful output, while using the output of the authors command for
example.