Report docker-compose ports in use
Along with reporting, defines a standard of sorts so reporting programs
can be written. Although duplication is happening, all scripts are
aiming to be used separately so one can copy just one script and use it
as expected. Not sure if it will work though.

Some common things to use in another scripts:

* DEBUG variable displays trace
* main script in main() function
* Help and version are displayed
* Option parsing
* Script header and vim config