Eases iteration on repositories
A common pattern to explore git repositories is to execute the same
command in all repositories inside a directory. Usually just one command
is needed, and if many are required a simple shell-script can do the
trick - and leave that "one liner" readable to others.

Scripts using the old "git-repositories" which just printed the name of
repositories for other scripts to iterate is dead. That functionality is
absorbed by "for_every_dir_containing" script now making it a weird
behaviour, sometimes you can pass an argument to that script which will
do nothing (be ignored) but directories will be print in the same way...
I probably want to remove that behavior.