Thinks about a talk on Git and version control
Thanks god. Using version control is not a debate between developers
anymore. Thanks to [CVS][], [Git][] and the popularization of the
development career. Now, we just need to use it better.

[CVS]: http://sources.vsta.org/comp.sources.unix/volume6/cvs/part1
[Git]: https://en.wikipedia.org/wiki/Git#History

What you do when you start in a new Company?
--------------------------------------------

Take a look in this repository for how much information you can retrieve
from a Git repository.

From where bugs spawn? (How to use Git)
---------------------------------------

Every code written is [bug][] in potential, changes to existing code
increase that potential. Bug fixes can create new bugs.

I treat bugs like monsters on an MMORPG, you can kill all of them
and yet do so again if you wait long enough.

[bug]: https://twitter.com/augustohp/status/1156201781396955137 "curl 31-bit bug"

> Allowing bugs to be found through valuable information

Fixing bugs is nasty, constant, business. Once you identify the root
cause of it, you could see all about when it was introduced:

    $ git log --patch -- src/file-with-bug.ext

Was the person aware of that problem while introducing the bug?
There are a multitude of different situations you may have questions
about but the most impactful is: Do I need to contact them to know?

No, you don't. Commit messages are there to provide those answers.
You just need to write them before they happen. Do you need to think
on every possible answer needed in the future? No. You can think about
the most probable ones:

* When changing a plethora of files at the same time using some automated
  tool, tell about it. Maybe even reproduce the command executed
* Explain about a refactor, its name and motivation
* What was the expected behavior before a bug is fixed?

If the answers (commits) for possible questions are broken into parts
hard to understand - either because they are too big, small, or contain
multiple changes together - changing things become more like a gamble.
People working in the same code base usually have to respect
constraints, like time, so solving a puzzle for every change is not a
recipe to have many changes made.

    $ git log --patch v1.0.0..v1.1.0 -- src/file-with-bug.ext

You can examine the history in many ways with `git log`, limiting it to
a _start_ and an _end_. Those references can be any Git reference, a
hash, a tag or a branch.

    $ git log --oneline v1.0.0.. -- src/file-with-bug.ext

It may be useful to read just the first lines of commit messages, just
to get a gist of the history until now.

> Diffs are a *very* portable way of sharing code

Any VCS is basically a bunch of commands that makes exchanging diffs
between files easier. Sometimes the requirements to execute these
commands are not present, like when logging in into production servers -
not recommended, and you are relieved to discover that those nasty diffs
are very easy to share: they are texts.

    $ git apply fixes-recover-my-password-url-template.patch

It is easy to apply patches using Git. You just need to have them, apply
them and them commit.

    $ git log --patch --raw v1.1.19.. > v1.1.20.patch

You can generate them too, and share with other people without sharing
your history and all sensitive information it may have.

    $ patch < v1.1.20.patch

You don't even need Git to apply them.

> When, what, who and where were the changes made?

Treat commits as chapters in a story. Any VCS will provide some
information on every "chapter": what is changing, who is changing and
when it is being changed. Writing them again is an effort that will
always result on having the same information again, in poorer form.

Provide contexts. Why it is being done the way it is? Are there
other solutions? Why they are not being used? Who this solution
impacts? Does this solution has an improvement to be done in the
future? Does it has other uses that we are not currently
exploring? What problems may arise? Who it will impact?
Does this needs any extra attention to monitor? What errors will
it produce and how should we fix them?

Use the imperative. Increase the immersion on a story where
what you read is being done as you read it. Like God on Genesis.

> Identification

Hopefully, you do not work with criminals. Hopefully your work has
direct impact on the lives of many people, which will require
identification of every people participating on that impact. It is as
simple as leaving your e-mail, and your name if you feeling awesome, on
every change you produce.

Just know that as you are free to type any e-mail, so is everyone.
Meaning that criminals would have an easy time making mistakes in your
name. Hopefully you can sign your changes, like putting a password
in all your changes.

Signing commits is useful because you leave the company. Your e-mail
will probably be deleted and the only way to prove your identity might
be your signature. ;)

> E-mail came before the internet, so has etiquette rules

* Always use your work, or most durable, e-mail
* First commit line is the subject, keep it under 50 chars
* Second line separates the *subject* from the *body*
* [Wrap body lines at 72 chars][1] and put a newline after paragraphs
* Use the same language from the other commit messages
* Don't commit [commented code][]
* Never break the build

[1]: https://github.com/torvalds/linux/pull/17#issuecomment-5661185 "word-wrapping is a property of the text"
[commented code]: https://twitter.com/lunetics/status/1155809252642033664 "Increase code coverage"

Conclusion
----------

If, someday, developers really get in a social network made for them,
commits will be the only acceptable form of interaction. They are so
rich, yet preventing distractions like stars, projects or likes.

Repositories could be awesome to read like books, interactive like
games, rich like HQs and free like in beer. And be the tool they are
meant to be: one to improve the quality of the software it holds.

PS: Isn't markdown beautiful? '#' are comments on commits, allow an
    alternate way to describe titles really is the work of a competent
    designer
PPS: Just though of a slogan to promote better commit messages: "I am no
     programmer, I am a committer" 🤷🏻‍♂️

Post-Title: Using Git
Post-Category: talk
Post-Tags: programming, version-control